Organize imports an add spaces and fix typos
