+++ src/artworks/client/actions/deleteArtworkAction.ts
@@ -1,13 +1,13 @@
-import artworksClient from "../../../client/ArtworksClient";
 import {
-  notify,
+  notifyDeleteArtworkSuccess,
   notifyDeleteArtworkError,
-} from "../../../toasts/deleteArtworkkToast/notify";
+} from "../../toasts/deleteArtworkToast/notify";
+import artworksClient from "../ArtworksClient";
 
 const deleteArtwork = async (artworkId: string): Promise<void> => {
   try {
     await artworksClient.deleteArtworkById(artworkId);
-    notify();
+    notifyDeleteArtworkSuccess();
   } catch (error) {
     notifyDeleteArtworkError(error as Error);
   }
