+++ src/artworks/client/actions/toggleArtworkFavouriteAction.ts
@@ -1,6 +1,6 @@
-import artworksClient from "../../../client/ArtworksClient";
-import { Artwork, ArtworkUpdate } from "../../../types";
-import { notifyToggleIsFavouriteArtworksError } from "../../../toasts/toggleArtworkIsfavourite/notify";
+import artworksClient from "../ArtworksClient";
+import { Artwork, ArtworkUpdate } from "../../types";
+import { notifyToggleIsFavouriteArtworksError } from "../../toasts/toggleArtworkIsfavourite/notify";
 
 const toggleArtworkFavourite = async (artwork: Artwork): Promise<Artwork> => {
   const updateData: ArtworkUpdate = {
@@ -9,9 +9,11 @@ const toggleArtworkFavourite = async (artwork: Artwork): Promise<Artwork> => {
   };
   try {
     const updatedArtwork = await artworksClient.updateArtwork(updateData);
+
     return updatedArtwork;
   } catch (error) {
     notifyToggleIsFavouriteArtworksError(error as Error);
+
     return artwork;
   }
 };
