+++ src/artworks/components/ArtworkDetail/ArtworkDetail.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { Artwork } from "../../types";
 import "./ArtworkDetail.scss";
-import { convertNumberto3DigitsString } from "../ArtworkCard/functions/convertNumberTo3DigitString";
 
 interface ArtworkDetailProps {
   artwork: Artwork;
@@ -16,8 +15,8 @@ const ArtworkDetail = ({ artwork }: ArtworkDetailProps): React.ReactElement => {
         loading="lazy"
         src={artwork.artworkUrl}
         alt={artwork.description}
-        width={convertNumberto3DigitsString(artwork.size.width)}
-        height={convertNumberto3DigitsString(artwork.size.height)}
+        width={artwork.size.width.toString()}
+        height={artwork.size.height.toString()}
       />
       <p className="artwork-detail__authorship">{`${artwork.author}, ${artwork.year}`}</p>
       <div className="artwork-detail__info">
