+++ src/artworks/pages/ArtworkFormPage/ArtworkFormPage.test.tsx
@@ -1,9 +1,9 @@
 import { render, screen } from "@testing-library/react";
-import ArtworkFormPage from "./ArtworkFormPage";
 import userEvent from "@testing-library/user-event";
 import { MemoryRouter } from "react-router-dom";
-import { server } from "../../../mocks/node";
+import ArtworkFormPage from "./ArtworkFormPage";
 import { http, HttpResponse } from "msw";
+import { server } from "../../../mocks/node";
 import routes from "../../../routes/routes";
 import { Bounce, ToastContainer } from "react-toastify";
 import { mockMonaLisa } from "../../mocks/artworks";
