+++ src/artworks/pages/ArtworkFormPage/ArtworkFormPage.tsx
@@ -1,21 +1,7 @@
 import React from "react";
 import ArtworkForm from "../../components/ArtworkForm/ArtworkForm";
-import { NewArtworkData } from "../../types";
-import artworksClient from "../../client/ArtworksClient";
-import {
-  notifyCreateArtworkError,
-  notify,
-} from "../../toasts/createArtworkToasts/notify";
+import createArtwork from "../../client/actions/createArtwork";
 import "./ArtworkFormPage.scss";
-
-const createArtwork = async (newArtworkData: NewArtworkData): Promise<void> => {
-  try {
-    await artworksClient.createArtwork(newArtworkData);
-    notify();
-  } catch (error) {
-    notifyCreateArtworkError(error as Error);
-  }
-};
 
 const ArtworkFormPage = (): React.ReactElement => {
   return (
