+++ src/artworks/toasts/deleteArtworkToast/DeleteArtworkSuccess.scss
