+++ src/artworks/toasts/deleteArtworkToast/DeleteArtworkSuccess.tsx
