+++ src/components/Loading/Loading.scss
@@ -5,7 +5,6 @@
   display: flex;
   justify-content: center;
   position: absolute;
-  background-color: #ffffffb1;
 
   &__content {
     position: fixed;
